feat(ws): Split the Connect button, such that clicking it opens the default (main) endpoint
