+++ workspaces/frontend/src/app/pages/Workspaces/Workspaces.tsx
@@ -33,9 +33,9 @@ import { ExpandedWorkspaceRow } from '~/app/pages/Workspaces/ExpandedWorkspaceRo
 import DeleteModal from '~/shared/components/DeleteModal';
 import { buildKindLogoDictionary } from '~/app/actions/WorkspaceKindsActions';
 import useWorkspaceKinds from '~/app/hooks/useWorkspaceKinds';
+import { WorkspaceConnectAction } from '~/app/pages/Workspaces/WorkspaceConnectAction';
 import Filter, { FilteredColumn } from 'shared/components/Filter';
 import { formatRam } from 'shared/utilities/WorkspaceResources';
-import { EndpointsDropdown } from '~/app/pages/Workspaces/WorkspaceEndpoints';
 
 export const Workspaces: React.FunctionComponent = () => {
   /* Mocked workspaces, to be removed after fetching info from backend */
@@ -71,8 +71,8 @@ export const Workspaces: React.FunctionComponent = () => {
         endpoints: [
           {
             displayName: 'JupyterLab',
-            port: '7777'
-          }
+            port: '7777',
+          },
         ],
       },
       options: {
@@ -122,11 +122,11 @@ export const Workspaces: React.FunctionComponent = () => {
         endpoints: [
           {
             displayName: 'JupyterLab',
-            port: '8888'
+            port: '8888',
           },
           {
             displayName: 'Spark Master',
-            port: '9999'
+            port: '9999',
           },
         ],
       },
@@ -479,7 +479,7 @@ export const Workspaces: React.FunctionComponent = () => {
                       </Timestamp>
                     </Td>
                     <Td>
-                      <EndpointsDropdown workspace={workspace}/>
+                      <WorkspaceConnectAction workspace={workspace} />
                     </Td>
                     <Td isActionCell data-testid="action-column">
                       <ActionsColumn
